修改内容：错误类型不一致修改

+++ lib/model/categoryGoodsList.dart
@@ -30,7 +30,7 @@ class CategoryGoodsListModel {
 class CategoryListData {
   String image;
   double oriPrice;
-  int presentPrice;
+  double presentPrice;
   String goodsName;
   String goodsId;
